+++ src/core/config/index.ts
@@ -1,0 +1,5 @@
+export * from './defaults';
+export * from './messages';
+export * from './services';
+export * from './icons';
+export * from './ui';